beta-signup-inbox/requirements.txt
flask>=3.0.0
flask-cors>=5.0.0
gunicorn>=21.0.0
sema-sdk>=0.1.0
resend>=2.0.0
google-genai>=1.0.0
boto3>=1.35.0
python-dotenv>=1.0.0
pytest>=8.0.0
